+++ app/src/main/java/com/bjz/imusicteacher/model/network/prediction/Prediction.java
@@ -6,7 +6,7 @@ package com.bjz.imusicteacher.model.network.prediction;
 
 public class Prediction {
     double[] probabilities;
-    double processingTime;
+    long processingTime;
 
     public double[] getProbabilities() {
         return probabilities;
@@ -17,11 +17,11 @@ public class Prediction {
         return this;
     }
 
-    public double getProcessingTime() {
+    public long getProcessingTime() {
         return processingTime;
     }
 
-    public Prediction setProcessingTime(double processingTime) {
+    public Prediction setProcessingTime(long processingTime) {
         this.processingTime = processingTime;
         return this;
     }
